Move tasks into rubberduke.tasks
