+++ src/main/java/rubberduke/Parser.java
@@ -1,5 +1,9 @@
 package rubberduke;
 
+import rubberduke.tasks.Deadline;
+import rubberduke.tasks.Event;
+import rubberduke.tasks.Todo;
+
 public class Parser {
     private final TaskList taskList;
